backend/requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.2.1
boto3==1.35.99
botocore==1.35.99
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
ecdsa==0.19.0
envparse==0.2.0
fastapi==0.115.6
fastapi-storages==0.3.0
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
orjson==3.10.13
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.2.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
redis==5.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
WTForms==3.1.2
yarl==1.17.1
